#67 原管理控制器、页面，移入sample作为展示
