--- app/controllers/kc_courses/application_controller.rb
@@ -1,10 +1,0 @@
-module KcCourses
-  class ApplicationController < ActionController::Base
-    layout "kc_courses/application"
-
-    if defined? PlayAuth
-      include PlayAuth::SessionsHelper
-      helper PlayAuth::SessionsHelper
-    end
-  end
-end
