--- app/controllers/kc_courses/home_controller.rb
@@ -1,9 +1,0 @@
-module KcCourses
-  class HomeController < KcCourses::ApplicationController
-    def index
-      @courses_count = KcCourses::Course.count
-      @chapters_count = KcCourses::Chapter.count
-      @wares_count = KcCourses::Ware.count
-    end
-  end
-end
