+++ sample/app/assets/javascripts/application.js
