+++ sample/app/assets/javascripts/uploader.coffee
