+++ sample/app/controllers/application_controller.rb
@@ -1,0 +1,6 @@
+class ApplicationController < ActionController::Base
+  if defined? PlayAuth
+    include PlayAuth::SessionsHelper
+    helper PlayAuth::SessionsHelper
+  end
+end
