+++ sample/app/views/chapters/index.html.haml
