+++ sample/app/views/chapters/new.html.haml
