+++ sample/app/views/chapters/show.html.haml
