+++ sample/app/views/courses/_form.html.haml
@@ -1,4 +1,4 @@
-= simple_form_for @course, wrapper: :kc_courses do |f|
+= simple_form_for @course, wrapper: :horizontal_form do |f|
   .form-inputs
     = f.input :title
     = f.input :desc, as: :text
