+++ sample/app/views/courses/edit.html.haml
