+++ sample/app/views/courses/index.html.haml
