+++ sample/app/views/courses/new.html.haml
