+++ sample/app/views/courses/show.html.haml
