+++ sample/app/views/home/index.html.haml
