+++ sample/app/views/joins/index.html.haml
