+++ sample/app/views/layouts/_navigation.html.haml
