+++ sample/app/views/layouts/application.html.haml
@@ -4,14 +4,14 @@
     %meta{:'http-equiv' => 'Content-Type', :content => 'text/html; charset=utf-8'}
     %meta{:name => 'viewport', :content => 'width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no'}
 
-    %title rails engine: kc_courses
+    %title rails engine: kc_courses sample
     = csrf_meta_tags
-    = stylesheet_link_tag 'kc_courses/application', :media => 'screen'
-    = javascript_include_tag 'kc_courses/application'
+    = stylesheet_link_tag 'application', :media => 'screen'
+    = javascript_include_tag 'application'
 
   %body
     %header
-      = render 'layouts/kc_courses/navigation'
+      = render 'layouts/navigation'
       
     %main{role: "main"}
       .container
