+++ sample/app/views/wares/_form.html.haml
@@ -1,4 +1,4 @@
-= simple_form_for [@chapter, @ware], wrapper: :kc_courses do |f|
+= simple_form_for [@chapter, @ware], wrapper: :horizontal_form do |f|
   .form-inputs
     = f.input :title
     = f.input :desc, as: :text
