+++ sample/app/views/wares/edit.html.haml
