+++ sample/app/views/wares/new.html.haml
